// src/gql/book/__generated__/Book.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: Book
// ====================================================

export interface Book {
  bookId: number;
  title: string;
  author: string;
  price: number;
}